Serve the externalized worker from the doenetml-iframe dev server
The externalized-worker change made the `@doenet/standalone` bundle load its
core worker from `/doenetml-worker/index.js`, but only `cypress.config.ts`
served that path. The plain `npm run dev` server (`vite.config.ts` +
`src/test-main.tsx`, which loads standalone from a Blob URL) never served it,
so the worker `importScripts` 404'd and the viewer never booted. CI stayed
green because it only runs the Cypress component tests, which use the config
that already served the worker.

Extract the worker-serving middleware into a shared `serveDoenetmlWorkerPlugin`
and use it from both the dev server and the Cypress config (removing the
duplicated inline middleware). Exclude the plugin from the package tsc build,
matching the other root-level config files.
